restructure RL so that user can customize environment
